+++ packages/react-native-sdk/src/components/Call/CallContent/CallContent.tsx
@@ -16,25 +16,39 @@ import {
 } from '../CallControls';
 import { useCall, useCallStateHooks } from '@stream-io/video-react-bindings';
 import { CallingState } from '@stream-io/video-client';
-import { useIncallManager } from '../../../hooks/useIncallManager';
-import { CallParticipantsListComponentProps } from '../CallParticipantsList';
-import { ParticipantViewComponentProps } from '../../Participant';
+import { useIncallManager } from '../../../hooks';
+import { Z_INDEX } from '../../../constants';
+import { useDebouncedValue } from '../../../utils/hooks';
+import {
+  FloatingParticipantView as DefaultFloatingParticipantView,
+  FloatingParticipantViewProps,
+  ParticipantViewComponentProps,
+} from '../../Participant';
+import { useTheme } from '../../../contexts';
 
-export type CallParticipantsComponentProps =
-  CallParticipantsListComponentProps &
-    Pick<
-      CallParticipantsGridProps,
-      'CallParticipantsList' | 'LocalParticipantView'
-    > & {
-      /**
-       * Component to customize the CallTopView component.
-       */
-      CallTopView?: React.ComponentType<CallTopViewProps> | null;
-      /**
-       * Component to customize the CallControls component.
-       */
-      CallControls?: React.ComponentType<CallControlProps> | null;
-    };
+export type CallParticipantsComponentProps = Pick<
+  CallParticipantsGridProps,
+  | 'CallParticipantsList'
+  | 'ParticipantLabel'
+  | 'ParticipantNetworkQualityIndicator'
+  | 'ParticipantReaction'
+  | 'ParticipantVideoFallback'
+  | 'ParticipantView'
+  | 'VideoRenderer'
+> & {
+  /**
+   * Component to customize the CallTopView component.
+   */
+  CallTopView?: React.ComponentType<CallTopViewProps> | null;
+  /**
+   * Component to customize the CallControls component.
+   */
+  CallControls?: React.ComponentType<CallControlProps> | null;
+  /**
+   * Component to customize the FloatingParticipantView.
+   */
+  FloatingParticipantView?: React.ComponentType<FloatingParticipantViewProps> | null;
+};
 
 export type CallContentProps = Pick<CallControlProps, 'onHangupCallHandler'> &
   Pick<
@@ -55,19 +69,33 @@ export const CallContent = ({
   CallParticipantsList,
   CallTopView = DefaultCallTopView,
   CallControls = DefaultCallControls,
+  FloatingParticipantView = DefaultFloatingParticipantView,
   ParticipantLabel,
   ParticipantNetworkQualityIndicator,
   ParticipantReaction,
   ParticipantVideoFallback,
   ParticipantView,
   ParticipantsInfoBadge,
-  LocalParticipantView,
   VideoRenderer,
   layout,
 }: CallContentProps) => {
+  const {
+    theme: { callContent },
+  } = useTheme();
   const { useHasOngoingScreenShare } = useCallStateHooks();
   const hasScreenShare = useHasOngoingScreenShare();
+  const { useRemoteParticipants, useLocalParticipant } = useCallStateHooks();
+
+  const _remoteParticipants = useRemoteParticipants();
+  const localParticipant = useLocalParticipant();
+  const remoteParticipants = useDebouncedValue(_remoteParticipants, 300); // we debounce the remote participants to avoid unnecessary rerenders that happen when participant tracks are all subscribed simultaneously
   const showSpotlightLayout = hasScreenShare || layout === 'spotlight';
+
+  const showFloatingView =
+    !showSpotlightLayout &&
+    remoteParticipants.length > 0 &&
+    remoteParticipants.length < 3;
+
   /**
    * This hook is used to handle IncallManager specs of the application.
    */
@@ -97,7 +125,6 @@ export const CallContent = ({
     ...participantViewProps,
     ParticipantView,
     CallParticipantsList,
-    LocalParticipantView,
   };
 
   const callParticipantsSpotlightProps: CallParticipantsSpotlightProps = {
@@ -107,21 +134,30 @@ export const CallContent = ({
   };
 
   return (
-    <View style={styles.container}>
-      <View style={styles.container}>
-        {CallTopView && (
-          <CallTopView
-            onBackPressed={onBackPressed}
-            onParticipantInfoPress={onParticipantInfoPress}
-            ParticipantsInfoBadge={ParticipantsInfoBadge}
-          />
-        )}
+    <View style={[styles.container, callContent.container]}>
+      <View style={[styles.container, callContent.callParticipantsContainer]}>
+        <View style={[styles.view, callContent.topContainer]}>
+          {CallTopView && (
+            <CallTopView
+              onBackPressed={onBackPressed}
+              onParticipantInfoPress={onParticipantInfoPress}
+              ParticipantsInfoBadge={ParticipantsInfoBadge}
+            />
+          )}
+          {showFloatingView && FloatingParticipantView && (
+            <FloatingParticipantView
+              participant={localParticipant}
+              {...participantViewProps}
+            />
+          )}
+        </View>
         {showSpotlightLayout ? (
           <CallParticipantsSpotlight {...callParticipantsSpotlightProps} />
         ) : (
           <CallParticipantsGrid {...callParticipantsGridProps} />
         )}
       </View>
+
       {CallControls && (
         <CallControls onHangupCallHandler={onHangupCallHandler} />
       )}
@@ -131,4 +167,8 @@ export const CallContent = ({
 
 const styles = StyleSheet.create({
   container: { flex: 1 },
+  view: {
+    ...StyleSheet.absoluteFillObject,
+    zIndex: Z_INDEX.IN_FRONT,
+  },
 });