+++ packages/react-native-sdk/src/components/Call/CallLayout/CallParticipantsGrid.tsx
@@ -8,20 +8,12 @@ import {
   CallParticipantsListComponentProps,
 } from '../CallParticipantsList/CallParticipantsList';
 import { ComponentTestIds } from '../../../constants/TestIds';
-import {
-  LocalParticipantView as DefaultLocalParticipantView,
-  LocalParticipantViewProps,
-} from '../../Participant';
-import { theme } from '../../../theme';
+import { useTheme } from '../../../contexts/ThemeContext';
 
 /**
  * Props for the CallParticipantsGrid component.
  */
 export type CallParticipantsGridProps = CallParticipantsListComponentProps & {
-  /**
-   * Component to customize the LocalParticipantView.
-   */
-  LocalParticipantView?: React.ComponentType<LocalParticipantViewProps> | null;
   /**
    * Component to customize the CallParticipantsList.
    */
@@ -33,7 +25,6 @@ export type CallParticipantsGridProps = CallParticipantsListComponentProps & {
  */
 export const CallParticipantsGrid = ({
   CallParticipantsList = DefaultCallParticipantsList,
-  LocalParticipantView = DefaultLocalParticipantView,
   ParticipantLabel,
   ParticipantNetworkQualityIndicator,
   ParticipantReaction,
@@ -41,6 +32,9 @@ export const CallParticipantsGrid = ({
   ParticipantView,
   VideoRenderer,
 }: CallParticipantsGridProps) => {
+  const {
+    theme: { colors, callParticipantsGrid },
+  } = useTheme();
   const { useRemoteParticipants, useParticipants } = useCallStateHooks();
   const _remoteParticipants = useRemoteParticipants();
   const allParticipants = useParticipants();
@@ -62,12 +56,13 @@ export const CallParticipantsGrid = ({
 
   return (
     <View
-      style={styles.container}
+      style={[
+        styles.container,
+        { backgroundColor: colors.dark_gray },
+        callParticipantsGrid.container,
+      ]}
       testID={ComponentTestIds.CALL_PARTICIPANTS_GRID}
     >
-      {showFloatingView && LocalParticipantView && (
-        <LocalParticipantView {...participantViewProps} />
-      )}
       {CallParticipantsList && (
         <CallParticipantsList
           participants={participants}
@@ -81,6 +76,5 @@ export const CallParticipantsGrid = ({
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: theme.light.dark_gray,
   },
 });