+++ packages/react-native-sdk/src/components/Participant/ParticipantView/ParticipantView.tsx
@@ -1,7 +1,6 @@
 import React, { ComponentType } from 'react';
 import { StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
 import { StreamVideoParticipant } from '@stream-io/video-client';
-import { theme } from '../../../theme';
 import {
   ParticipantNetworkQualityIndicator as DefaultParticipantNetworkQualityIndicator,
   ParticipantNetworkQualityIndicatorProps,
@@ -22,6 +21,7 @@ import {
   VideoRenderer as DefaultVideoRenderer,
   VideoRendererProps,
 } from './VideoRenderer';
+import { useTheme } from '../../../contexts/ThemeContext';
 
 export type ParticipantVideoType = 'video' | 'screen';
 
@@ -68,7 +68,7 @@ export type ParticipantViewProps = ParticipantViewComponentProps & {
   /**
    * The video kind that will be displayed.
    */
-  videoMode: ParticipantVideoType;
+  videoMode?: ParticipantVideoType;
   /**
    * Custom style to be merged with the participant view.
    */
@@ -88,7 +88,7 @@ export type ParticipantViewProps = ParticipantViewComponentProps & {
  */
 export const ParticipantView = ({
   participant,
-  videoMode,
+  videoMode = 'video',
   isVisible = true,
   style,
   ParticipantLabel = DefaultParticipantLabel,
@@ -98,10 +98,19 @@ export const ParticipantView = ({
   ParticipantVideoFallback = DefaultParticipantVideoFallback,
   videoZOrder = 0,
 }: ParticipantViewProps) => {
+  const {
+    theme: { colors, participantView },
+  } = useTheme();
   const { isSpeaking, userId } = participant;
   const isScreenSharing = videoMode === 'screen';
   const applySpeakerStyle = isSpeaking && !isScreenSharing;
-  const speakerStyle = applySpeakerStyle && styles.isSpeaking;
+  const speakerStyle = applySpeakerStyle && [
+    styles.highligtedContainer,
+    {
+      borderColor: colors.primary,
+    },
+    participantView.highligtedContainer,
+  ];
 
   return (
     <View
@@ -122,7 +131,7 @@ export const ParticipantView = ({
           videoZOrder={videoZOrder}
         />
       )}
-      <View style={styles.bottomView}>
+      <View style={[styles.footerContainer, participantView.footerContainer]}>
         {ParticipantLabel && (
           <ParticipantLabel participant={participant} videoMode={videoMode} />
         )}
@@ -137,17 +146,16 @@ export const ParticipantView = ({
 const styles = StyleSheet.create({
   container: {
     justifyContent: 'space-between',
-    padding: theme.padding.xs,
+    padding: 4,
     overflow: 'hidden',
     margin: 2,
   },
-  bottomView: {
+  footerContainer: {
     flexDirection: 'row',
     justifyContent: 'space-between',
     alignItems: 'center',
   },
-  isSpeaking: {
-    borderColor: theme.light.primary,
+  highligtedContainer: {
     borderWidth: 2,
   },
 });