+++ packages/react-native-sdk/src/theme/button.ts
@@ -1,9 +1,0 @@
-import { ButtonScheme } from './types';
-
-export const button: ButtonScheme = {
-  xs: { height: 40, width: 40, borderRadius: 20 },
-  sm: { height: 50, width: 50, borderRadius: 25 },
-  md: { height: 60, width: 60, borderRadius: 30 },
-  lg: { height: 70, width: 70, borderRadius: 35 },
-  xl: { height: 80, width: 80, borderRadius: 40 },
-};