+++ packages/react-native-sdk/src/theme/rounded.ts
@@ -1,9 +1,0 @@
-import { RoundedScheme } from './types';
-
-export const rounded: RoundedScheme = {
-  xs: 5,
-  sm: 10,
-  md: 15,
-  lg: 20,
-  xl: 30,
-};