+++ sample-apps/react/egress-composite/src/components/layouts/PaginatedGrid/PaginatedGrid.tsx
@@ -1,0 +1,29 @@
+import {
+  DefaultParticipantViewUI,
+  PaginatedGridLayout,
+} from '@stream-io/video-react-sdk';
+import { useConfigurationContext } from '../../../ConfigurationContext';
+
+import './PaginatedGrid.scss';
+
+export const PaginatedGrid = () => {
+  // const setParticipantVideoRef = useEgressReadyWhenAnyParticipantMounts();
+  const {
+    options: { 'layout.grid.page_size': pageSize = 20 },
+  } = useConfigurationContext();
+
+  return (
+    <div className="paginated-grid">
+      <PaginatedGridLayout
+        ParticipantViewUI={
+          <DefaultParticipantViewUI
+            indicatorsVisible={false}
+            showMenuButton={false}
+          />
+        }
+        excludeLocalParticipant
+        groupSize={pageSize}
+      />
+    </div>
+  );
+};