+++ sample-apps/react/react-video-demo/src/components/InvitePanel/InvitePanel.tsx
@@ -5,7 +5,7 @@ import classnames from 'classnames';
 import { Copy, Reload, UserChecked } from '../Icons';
 import Panel from '../Panel';
 
-import { useUserContext } from '../../contexts/UserContext';
+import { usePanelContext, PANEL_VISIBILITY } from '../../contexts/PanelContext';
 
 import { useBreakpoint } from '../../hooks/useBreakpoints';
 
@@ -84,15 +84,20 @@ export const InvitePanel = ({
   isFocused,
   callId,
 }: InvitePanelProps) => {
-  const [showQr, setShowQr] = useState(false);
-  const { qr } = useUserContext();
+  const [showQr, setShowQr] = useState(true);
   const breakpoint = useBreakpoint();
 
+  const { qrCodeVisibility } = usePanelContext();
+
   useEffect(() => {
-    if (breakpoint === 'lg') {
-      setShowQr(true);
+    if (qrCodeVisibility === PANEL_VISIBILITY.collapsed) {
+      setShowQr(false);
+    } else {
+      if (breakpoint === 'lg') {
+        setShowQr(true);
+      }
     }
-  }, [breakpoint]);
+  }, [breakpoint, qrCodeVisibility]);
 
   const handleToggleDisplayQr = useCallback(() => {
     setShowQr((prev) => !prev);
@@ -110,7 +115,6 @@ export const InvitePanel = ({
 
   const qrCodeContent = new URL(window.location.toString());
   qrCodeContent.searchParams.set('id', callId);
-  qrCodeContent.searchParams.set('qr', qr || 'true');
 
   return (
     <Panel
@@ -120,22 +124,14 @@ export const InvitePanel = ({
     >
       <>
         <Invite callId={callId} canShare={false} />
+        <p className={styles.description} onClick={handleToggleDisplayQr}>
+          Or scan the QR code with your phone to test it yourself:
+          <span className={showQrIndicatorClassNames}>▼</span>
+        </p>
 
-        {Boolean(qr) ? (
-          <>
-            <p className={styles.description} onClick={handleToggleDisplayQr}>
-              Or scan the QR code with your phone to test it yourself:
-              <span className={showQrIndicatorClassNames}>▼</span>
-            </p>
-
-            <div className={qrClassNames}>
-              <QRCodeSVG
-                className={styles.code}
-                value={qrCodeContent.toString()}
-              />
-            </div>
-          </>
-        ) : null}
+        <div className={qrClassNames}>
+          <QRCodeSVG className={styles.code} value={qrCodeContent.toString()} />
+        </div>
       </>
     </Panel>
   );